user changes

+++ src/main/java/com/uber/booking/controllers/DriverRideController.java
@@ -29,18 +29,18 @@ public class DriverRideController {
 
 
     @PostMapping(value = "/bookRide")
-    public BookingDetailsResponse bookRide(@NotNull @Param("rideId") Integer rideId, @Param("driverId") Integer driverId) throws IOException {
+    public BookingDetailsResponse bookRide(@NotNull @Param("rideId") Long rideId, @Param("driverId") Integer driverId) throws IOException {
         return driverRideService.bookRide(rideId, driverId);
     }
 
 
     @PutMapping(value = "/cancelRide")
-    public BookingDetailsResponse cancelRide(@Param("rideId") Integer rideId, @Param("driverId") Integer driverId) throws IOException {
+    public BookingDetailsResponse cancelRide(@Param("rideId") Long rideId, @Param("driverId") Integer driverId) throws IOException {
         return driverRideService.cancelRide(rideId, driverId);
     }
 
     @PutMapping(value = "/startRide")
-    public void startRide(@Param("rideId") Integer rideId, @Param("driverId") Integer driverId) throws IOException {
+    public void startRide(@Param("rideId") Long rideId, @Param("driverId") Integer driverId) throws IOException {
         driverRideService.startRide(rideId, driverId);
     }
 
@@ -50,7 +50,7 @@ public class DriverRideController {
      * @throws IOException
      */
     @PutMapping(value = "/endRide")
-    public void endRide(@Param("rideId") Integer rideId, @Param("driverId") Integer driverId) throws IOException {
+    public void endRide(@Param("rideId") Long rideId, @Param("driverId") Integer driverId) throws IOException {
         driverRideService.endRide(rideId, driverId);
     }
